Change the chkio return values to be size_t instead of ints. This can make a difference for large reads and writes of > 2GB files.
git-svn-id: file://localhost/data1/git/presto_svn/trunk@1358 5fac236c-a0d8-0310-9745-ef116914eda1

+++ include/chkio.h
@@ -13,17 +13,17 @@
 FILE *chkfopen(char *path, const char *mode);
 /* Preform a file open with error checking.  */
 
-int chkfread(void *data, size_t type, size_t number, FILE * stream);
+size_t chkfread(void *data, size_t type, size_t number, FILE * stream);
 /* Preform a file read with error checking.  */
 
-int chkfwrite(void *data, size_t type, size_t number, FILE * stream);
+size_t chkfwrite(void *data, size_t type, size_t number, FILE * stream);
 /* Preform a file write with error checking. */
 
-int chkfseek(FILE * stream, long offset, int whence);
+size_t chkfseek(FILE * stream, long offset, int whence);
 /* NOTE:  This is meant only for backwards compatibility.  */
 /* You should probably be calling chkfileseek() directly.  */
 
-int chkfileseek(FILE * stream, long offset, size_t size, int whence);
+size_t chkfileseek(FILE * stream, long offset, size_t size, int whence);
 /* Preform a file seek with error checking.  */
 
 long long chkfilelen(FILE *file, size_t size);
